bot/requirements.txt
requests==2.31.0
requests-oauthlib==1.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0